# swagger.yml
openapi: 3.0.2
info:
  title: Highway API
  description: API to recover highway
  version: "1.0.0"
  contact:
    email: koke106@gmail.com 
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: http://localhost:8000
    description: Local testing API
    
paths:
  /profile:
    get:
      summary: Get user profile
      description: This operation returns the profile information of the authenticated user.
      tags:
        - Perfil
      parameters:
        - name: Email
          in: header
          required: true
          description: User email
          schema:
            type: string
      responses:
        200:
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        401:
          description: Not authorized
        404:
          description: User not found
        500:
          description: Internal error
  /team:
    get:
      summary: Get team info
      description: This operation returns the info of the team, members, cars...
      tags:
        - Team
      parameters:
        - name: id
          in: header
          required: true
          description: Id of the team
          schema:
            type: string
      responses:
        200:
          description: Team info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        401:
          description: Not authorized
        404:
          description: User not found
        500:
          description: Internal error
  /car:
    get:
      summary: Get car info
      description: This operation returns all the car info, parts...
      tags:
        - Car
      parameters:
        - name: id
          in: header
          required: true
          description: Id of the car
          schema:
            type: string
      responses:
        200:
          description: Car info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        401:
          description: Not authorized
        404:
          description: User not found
        500:
          description: Internal error
  /assembly:
    get:
      summary: Get assembly info
      description: This operation return all assembly info of the car
      tags:
        - Assembly
      parameters:
        - name: id
          in: header
          required: true
          description: Id of the assembly
          schema:
            type: string
      responses:
        200:
          description: Assembly info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        401:
          description: Not authorized
        404:
          description: User not found
        500:
          description: Internal error
  /parts:
    get:
      summary: Parts
      description: This operation returns all available parts
      tags:
        - Parts
      responses:
        200:
          description: List of parts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/News'
        401:
          description: Not authorized
        404:
          description: User not found
        500:
          description: Internal error
  /news:
    get:
      summary: News
      description: This operation returns all the news
      tags:
        - News
      responses:
        200:
          description: List of news
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/News'
        401:
          description: Not authorized
        404:
          description: User not found
        500:
          description: Internal error

components:
  schemas:
    Profile:
      type: object
      properties:
        id:
          type: integer
          description: Unique ID user
        username:
          type: string
          description: Username
        name:
          type: string
          description: Name
        surname:
          type: string
          description: Surname
        email:
          type: string
          description: Email
        password:
          type: string
          description: Password
        rol:
          type: string
          description: User rol
        creation_date:
          type: string
          format: date-time
          description: Date creation
    Team:
      type: object
      properties: 
        id:
          type: integer
          description: Unique ID team
        id_profile:
          type: integer
          description: Unique ID profile
        name:
          type: string
          description: Team Name
    Car:
      type: object
      properties: 
        id:
          type: integer
          description: Unique ID car
        id_team:
          type: integer
          description: Unique ID team
        name:
          type: string
          description: Car Name
        brand:
          type: string
          description: Car Brand
        model:
          type: string
          description: Car Model
        year:
          type: string
          format: date-time
          description: Car Year
    Assembly:
      type: object
      properties:
        id:
          type: integer
          description: Unique ID assembly
        id_car:
          type: integer
          description: Unique ID car
        id_parts:
          type: integer
          description: Unique ID parts
        characteristics:
          type: string
          description: characteristics
    Parts:
      type: object
      properties:
        id:
          type: integer
          description: Unique ID parts
        name:
          type: string
          description: Part name
        description:
          type: integer
          description: Part description
    News:
      type: object
      properties:
        id:
          type: integer
          description: Unique ID new
        title:
          type: string
          description: News title
        content:
          type: string
          description: News content
        date:
          type: string
          format: date-time
          description: Date of the new